Reorg graph docs
